feat:调整钩子函数
